+++ src/ui-kit/experimental/patterns/layout/components/index.ts
@@ -2,7 +2,6 @@ export * from './actionbar.component';
 export * from './layout.component';
 export * from './main.component';
 export * from './pagination';
-export * from './toolbar.component';
 export * from './layout-components.module';
 export * from './resources-wrapper.component';
 export * from './title-section.component';
@@ -10,4 +9,3 @@ export * from './layout-components.module';
 export * from './page';
 export * from './main-content.component';
 export * from './content-outlet.directive';
-export * from './aside-toggle.component';
