+++ src/ui-kit/layout/aside-toggle/index.ts
@@ -1,0 +1,2 @@
+export { SamAsideToggleComponent } from './aside-toggle.component';
+export * from './aside-toggle.module';
