+++ src/ui-kit/layout/index.ts
@@ -1,2 +1,4 @@
+export * from './aside-toggle';
 export * from './filters-wrapper';
+export * from './toolbar';
 export * from './layout.module';
