+++ src/ui-kit/layout/layout.module.ts
@@ -1,8 +1,18 @@
 import { NgModule } from '@angular/core';
+import { SamAsideToggleModule } from './aside-toggle';
 import { SamFiltersWrapperModule } from './filters-wrapper';
+import { SamToolbarsModule } from './toolbar';
 
 @NgModule({
-    imports: [SamFiltersWrapperModule],
-    exports: [ SamFiltersWrapperModule ]
+    imports: [
+        SamAsideToggleModule,
+        SamFiltersWrapperModule,
+        SamToolbarsModule,
+    ],
+    exports: [
+        SamAsideToggleModule,
+        SamFiltersWrapperModule,
+        SamToolbarsModule,
+    ]
 })
 export class SamLayoutNextModule { }
