+++ src/ui-kit/layout/toolbar/index.ts
@@ -1,0 +1,2 @@
+export * from './toolbar.module';
+export { SamToolbarComponent } from './toolbar.component';
